Add files via upload
INTERNAL MEMO
Our Concept of Tokenization of Timeshare Through Non-Fungible Tokens (NFT) Ecosystem
INNOU, April 09, 2021
M. Whilst Play-Sell model is a working horse which will generate cash as in good times so in bad times (see White Paper on Play-Sell Collaborative Purchasing), we are currently (April 2021) approaching a perfect storm opportunity with tokenizing timeshare industry.
On the one hand, the classic timeshare industry is currently experiencing huge problems as the holders of timeshares realized, due to pandemic, that there is no secondary market for their contracts and the only way to get their money back is to sue the operators,
On the other hand, the Non-Fungible Tokens come into fashion which will last for the next year or so.
The first problem can be easily solved by tokenizing timeshare contracts through NFT and creating robust, solid secondary market for it.
It is just a matter of time when the largest operators will realize that, not they have not yet get this “Aha!” moment.
The only way we can do this is to get in touch with the largest operators directly. We need good contacts, either of our own, or through reliable intermediaries.
The technical implementation is very doable.