Add files via upload
INTERNAL MEMO
INNOU Investment Policy Principles in Tokenizing Real Estate
INNOU, March 21, 2021
M.
2
Digital tokenization of assets based on cryptographic distributed computation is changing the course and manner of business globally.
Application of these technologies are increasing the efficiencies of financing and capital formation, and, when thoughtfully applied, is fundamentally altering traditional forms and terms of fund raising.
Asset tokenization for raising funds is the carefully structured process whereby assets or liabilities are packaged, underwritten, and sold in the various forms of digital cryptographically protected and specifically distributed tokens which currently become commonly known as asset-backed tokens. Such tokens are essentially digitally created contracts with a set of self-governing rules and conditions.
As such, it is a subset of a broader trend, the general phenomenon whereby more and more fund raising is occurring through the agency of digital tokens.
Currently, there are forming three types of Real Estate-Backed Token1s for raising funds on the markets:
• Utility Tokens;
• Equity Tokens;
• Debt Tokens.
All three types of Tokens can have strong similarities to the degree of blending, depending on the way the structuring had been approached, which is especially picturesque for the first two types - Utility Tokens and Equity Tokens – due to the very nature of the real estate being used as underlying asset.
These asset-backed financing instruments normally have credit risk which can and must be structured.
Whilst INNOU is working on all three types of the tokens, we have distinctly strong focus on the Debt Tokens.
This memo provides a theoretical overview of asset-backed tokenization and a platform upon which to build an understanding and represents a basic tool kit – the working knowledge one needs to understand the structure of an asset-backed tokenization transactions.